requirements.txt
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorthief==0.2.1
comtypes==1.3.0
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.4
cssselect==1.2.0
cycler==0.12.1
dominate==2.9.1
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
fonttools==4.49.0
future==1.0.0
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
keyboard==0.13.5
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
MouseInfo==0.1.3
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parsel==1.8.1
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.10.4
pillow==10.2.0
platformdirs==4.2.0
Protego==0.3.0
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyDispatcher==2.0.7
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==24.0.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.27.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywin32==306
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
selenium==4.18.1
service-identity==24.1.0
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
tldextract==5.1.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==24.2.0rc1
twisted-iocpsupport==1.0.4
typing_extensions==4.10.0rc1
tzdata==2024.1
urllib3==2.2.1
visitor==0.1.3
w3lib==2.1.2
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
zope.interface==6.2
